customerapi/requirements.txt
Flask==2.3.2
pymongo==4.2.0
marshmallow==3.20.1
PyJWT==2.1.0